replace passport sessionID with SAML ID
